+++ src/onion_service.rs
@@ -577,7 +577,7 @@ fn generate_config_map(
 ) -> ConfigMap {
     ConfigMap {
         metadata: ObjectMeta {
-            name: Some(object.config_map_name().to_string()),
+            name: Some(object.config_map_name().into()),
             annotations: Some(annotations.into()),
             labels: Some(labels.into()),
             owner_references: Some(vec![object.controller_owner_ref(&()).unwrap()]),
@@ -604,7 +604,7 @@ fn generate_deployment(
 ) -> Deployment {
     Deployment {
         metadata: ObjectMeta {
-            name: Some(object.deployment_name().to_string()),
+            name: Some(object.deployment_name().into()),
             annotations: Some(annotations.into()),
             labels: Some(labels.into()),
             owner_references: Some(vec![object.controller_owner_ref(&()).unwrap()]),
@@ -716,7 +716,7 @@ fn generate_deployment(
                                     },
                                 ]),
                                 optional: Some(false),
-                                secret_name: Some(onion_key.secret_name().to_string()),
+                                secret_name: Some(onion_key.secret_name().into()),
                             }),
                             ..Default::default()
                         },
@@ -739,7 +739,7 @@ fn generate_deployment(
                                     }
                                     items
                                 }),
-                                name: Some(object.config_map_name().to_string()),
+                                name: Some(object.config_map_name().into()),
                                 optional: Some(false),
                             }),
                             ..Default::default()